relocated and renamed condition specifications to be inside a 'Badges' namespace
