+++ app/specifications/badges/abstract_condition_specification.rb
@@ -1,0 +1,12 @@
+module Badges
+  class Badges::AbstractConditionSpecification
+    def initialize(session, parameter)
+      @test_taking_session = session
+      @parameter = parameter
+    end
+
+    def satisfies?
+      raise "satisfies? is not implemented in the #{self.class}"
+    end
+  end
+end
